+++ resources/views/components/button/group.blade.php
@@ -1,0 +1,10 @@
+@props([
+    'vertical' => false,
+])
+<div {{ $attributes
+    ->class([
+        $vertical ? 'btn-group-vertical' : 'btn-group',
+    ])
+    ->merge([
+        'role' => 'group'
+    ]) }}>{{ $slot }}</div>