+++ resources/views/components/nav/index.blade.php
@@ -1,0 +1,7 @@
+@props([
+    /** @var string $container */
+    'container' => 'ul',
+])
+<{{ $container }} {{ $attributes->class('nav') }}>
+    {{ $slot }}
+</{{ $container }}>