+++ CHANGELOG.md
@@ -10,3 +10,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Components for forms and their fields including feedback
+- Components for breadcrumb, links, navigation
+- Components for buttons, groups, toolbars